capability tests
